src/test/java/io/novaordis/events/processing/TestProcedure.java
/*
 * Copyright (c) 2017 Nova Ordis LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package io.novaordis.events.processing;

import io.novaordis.events.api.event.Event;

import java.util.ArrayList;
import java.util.Collections;
import java.util.List;
import java.util.concurrent.atomic.AtomicLong;

/**
 * We introduced this to insure EndOfStream propagates to ProcedureBase subclasses.
 *
 * @author Ovidiu Feodorov <ovidiu@novaordis.com>
 * @since 8/10/17
 */
public class TestProcedure extends ProcedureBase {

    // Constants -------------------------------------------------------------------------------------------------------

    // Static ----------------------------------------------------------------------------------------------------------

    // Attributes ------------------------------------------------------------------------------------------------------

    private List<Event> received;

    // Constructors ----------------------------------------------------------------------------------------------------

    public TestProcedure() {

        this.received = new ArrayList<>();
    }

    // ProcedureBase overrides -----------------------------------------------------------------------------------------

    @Override
    protected void process(AtomicLong invocationCount, Event e) throws EventProcessingException {

        received.add(e);
    }

    @Override
    public List<String> getCommandLineLabels() {

        return Collections.singletonList("test");
    }

    // Public ----------------------------------------------------------------------------------------------------------

    public List<Event> getReceived() {

        return received;
    }

    // Package protected -----------------------------------------------------------------------------------------------

    // Protected -------------------------------------------------------------------------------------------------------

    // Private ---------------------------------------------------------------------------------------------------------

    // Inner classes ---------------------------------------------------------------------------------------------------

}
